requirements.txt
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
Deprecated==1.2.12
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
jsons==1.4.2
pycparser==2.20
PyDriller==2.0
pygit2==1.6.1
PyGithub==1.55
Pygments==2.9.0
PyJWT==2.1.0
PyNaCl==1.4.0
requests==2.26.0
rich==10.6.0
six==1.16.0
smmap==4.0.0
tqdm==4.61.2
typish==1.9.2
urllib3==1.26.6
wrapt==1.12.1
